# schema/openapi.yaml
openapi: "3.0.3"
info:
  title: Slimeopolis API
  description: |
    Global Slime Emporium API — Cloudflare API Shield demonstration.
    Upload this schema to: Security → API Shield → Schema Validation.
  version: "1.0.0"
  contact:
    name: Slimeopolis API
    url: https://slimeopolis.example.com
servers:
  - url: https://slimeopolis.example.com/api
    description: Production (replace with your actual domain)

# ─── Security Schemes ───────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /auth/login or /auth/register.
        API Shield Session Identifier: configure the Authorization header
        as the session identifier for Volumetric Abuse Detection.
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          minLength: 8
          example: mysecretpassword
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          minLength: 8
          example: mysecretpassword
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Jane Doe

    # ── Cart ────────────────────────────────────────────────────────────────────
    AddCartItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
          example: prod_jp_001
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          example: 2
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          example: 3

    # ── Orders ──────────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a new user
      description: |
        API Shield: Rate Limiting target (5 req/min per IP).
        Schema Validation: email, password (min 8 chars), name.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
        "409":
          description: Email already in use
        "422":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      summary: Sign in to get a JWT token
      description: |
        API Shield: PRIMARY rate limiting target (5 req/min per IP brute-force prevention).
        Volumetric Abuse Detection: per-session adaptive threshold monitoring.
        Schema Validation: email format + password presence.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful, returns JWT
        "401":
          description: Invalid credentials
        "422":
          description: Validation error
  /auth/logout:
    post:
      operationId: logoutUser
      summary: Sign out and revoke current token
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current user info
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user object

  # ── Products ──────────────────────────────────────────────────────────────────
...
  /cart:
    get:
      operationId: getCart
      summary: Get current user's cart
      description: |
        API Shield: Sequence Analytics step 3 (cart view).
      tags: [Cart]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cart contents with line items

    delete:
      operationId: clearCart
      summary: Clear all items from cart
      tags: [Cart]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cart cleared
  /cart/items:
    post:
      operationId: addCartItem
      summary: Add an item to cart
      description: |
        API Shield: Sequence Analytics step 4 (add to cart).
        Schema Validation: product_id string, quantity integer min 1.
      tags: [Cart]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Item quantity updated (if already in cart)
        "201":
          description: Item added to cart
        "422":
          description: Validation error or insufficient stock
  /cart/items/{itemId}:
    put:
      operationId: updateCartItem
      summary: Update quantity of a cart item
      description: |
        API Shield: Schema Validation — quantity integer 1-99.
      tags: [Cart]
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Cart item updated

    delete:
      operationId: removeCartItem
      summary: Remove an item from cart
      tags: [Cart]
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Item removed

  # ── Orders ────────────────────────────────────────────────────────────────────
...